gdbstub: implement a basic gdb remote serial protocol.
This configures the UART on the raspberry pi to be a remote serial
protocol.  This is good for low-level programming where you want to be
able to debug a system but don't have access to a JTAG.  It's pretty
basic - no watchpoints and there isn't true single stepping.

Known limitations:
  - single stepping works by setting breakpoints - and IRQ that fires in
    the middle of a step will be processed fully.
  - the rsp uses fiqs for handling serial input - the application can't
    use fiqs.
  - trustzone's secure monitor mode is used for handling breakpoints -
    the application can't use trustzone.
  - no watchpoints.